# API/Documentation/custom-product-controller.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WAX API - Custom Products",
...
    "version": "1.0.0",
    "contact": {
      "name": "WAX Backend Team"
    }
  },
  "servers": [
    {
      "url": "https://localhost:5007",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/CustomProduct/me": {
      "get": {
        "tags": ["CustomProduct - Customer"],
        "summary": "List my custom products",
        "description": "Returns all custom products belonging to the authenticated user, in any status. Useful for the customer's dashboard to track the negotiation state of each piece.",
        "operationId": "CustomProduct_GetMine",
        "responses": {
          "200": {
            "description": "List of custom products for the authenticated user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CustomProductDto" }
                },
                "example": [
                  {
                    "id": "cp-uuid-001",
                    "name": "Ring",
                    "description": "A custom engagement ring",
                    "price": 6000,
                    "taskId": "meshy-task-abc123",
                    "glbUrl": "https://cdn.meshy.ai/models/abc123.glb",
                    "ownerUserId": "user-uuid-001",
                    "status": "AwaitingCustomerReview",
                    "agreedPrice": null,
                    "design": {
                      "type": "Ring",
                      "material": "Resin",
                      "color": "Blue",
                      "shape": "Round",
                      "dimensions": "18x18x10",
                      "details": null
                    },
                    "proposals": [
                      {
                        "id": "proposal-uuid-001",
                        "amount": 7500,
                        "source": "System",
                        "comment": null,
                        "isAccepted": false,
                        "createdAt": "2026-05-10T14:00:00Z"
                      },
                      {
                        "id": "proposal-uuid-002",
                        "amount": 6000,
                        "source": "Admin",
                        "comment": "We can offer this piece at a lower price given the material.",
                        "isAccepted": false,
                        "createdAt": "2026-05-10T15:00:00Z"
                      }
                    ],
                    "createdAt": "2026-05-10T14:00:00Z",
                    "updatedAt": "2026-05-10T15:00:00Z"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "User is not authenticated"
          }
        },
        "security": [{ "cookieAuth": [] }]
      }
    },
    "/api/CustomProduct/{id}": {
      "get": {
        "tags": ["CustomProduct - Customer"],
        "summary": "Get custom product details",
        "description": "Returns the full details of a specific custom product including all price proposals. Customers can only access their own products. Admins can access any product.",
        "operationId": "CustomProduct_GetDetails",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique identifier of the custom product",
            "required": true,
            "schema": { "type": "string" },
            "example": "cp-uuid-001"
          }
        ],
        "responses": {
          "200": {
            "description": "Custom product details retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CustomProductDto" },
                "example": {
                  "id": "cp-uuid-001",
                  "name": "Ring",
                  "description": "A custom engagement ring with a blue gemstone",
                  "price": 6000,
                  "taskId": "meshy-task-abc123",
                  "glbUrl": "https://cdn.meshy.ai/models/abc123.glb",
                  "ownerUserId": "user-uuid-001",
                  "status": "AwaitingCustomerReview",
                  "agreedPrice": null,
                  "design": {
                    "type": "Ring",
                    "material": "Resin",
                    "color": "Blue",
                    "shape": "Round",
                    "dimensions": "18x18x10",
                    "details": "Engraved with initials JD"
                  },
                  "proposals": [
                    {
                      "id": "proposal-uuid-001",
                      "amount": 7500,
                      "source": "System",
                      "comment": null,
                      "isAccepted": false,
                      "createdAt": "2026-05-10T14:00:00Z"
                    },
                    {
                      "id": "proposal-uuid-002",
                      "amount": 6000,
                      "source": "Admin",
                      "comment": "We can offer this piece at a lower price given the material.",
                      "isAccepted": false,
                      "createdAt": "2026-05-10T15:00:00Z"
                    }
                  ],
                  "createdAt": "2026-05-10T14:00:00Z",
                  "updatedAt": "2026-05-10T15:00:00Z"
                }
              }
            }
          },
          "401": { "description": "User is not authenticated" },
          "403": { "description": "Customer trying to access another user's product" },
          "404": { "description": "Custom product not found" }
        },
        "security": [{ "cookieAuth": [] }]
      }
    },
    "/api/CustomProduct/admin": {
      "get": {
        "tags": ["CustomProduct - Admin"],
        "summary": "List all custom products (admin)",
        "description": "Returns all custom products across all users. Supports optional filtering by status. Intended for the admin panel to manage the negotiation queue.",
        "operationId": "CustomProduct_GetAdmin",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Optional status filter. When omitted, returns products in all statuses.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "PendingQuotation",
                "AwaitingAdminReview",
                "AwaitingCustomerReview",
                "Approved",
                "Rejected",
                "AddedToBasket"
              ]
            },
            "example": "AwaitingAdminReview"
          }
        ],
        "responses": {
          "200": {
            "description": "List of custom products matching the filter",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CustomProductDto" }
                },
                "example": [
                  {
                    "id": "cp-uuid-001",
                    "name": "Ring",
                    "description": "A custom engagement ring",
                    "price": 7500,
                    "taskId": "meshy-task-abc123",
                    "glbUrl": "https://cdn.meshy.ai/models/abc123.glb",
                    "ownerUserId": "user-uuid-001",
                    "status": "AwaitingAdminReview",
                    "agreedPrice": null,
                    "design": {
                      "type": "Ring",
                      "material": "Resin",
                      "color": "Blue",
                      "shape": "Round",
                      "dimensions": "18x18x10",
                      "details": null
                    },
                    "proposals": [
                      {
                        "id": "proposal-uuid-001",
                        "amount": 7500,
                        "source": "System",
                        "comment": null,
                        "isAccepted": false,
                        "createdAt": "2026-05-10T14:00:00Z"
                      }
                    ],
                    "createdAt": "2026-05-10T14:00:00Z",
                    "updatedAt": null
                  }
                ]
              }
            }
          },
          "401": { "description": "User is not authenticated" },
          "403": { "description": "User does not have the `Admin` role" }
        },
        "security": [{ "cookieAuth": [] }]
      }
    },
    "/api/CustomProduct/{id}/admin/proposals": {
      "post": {
        "tags": ["CustomProduct - Admin"],
        "summary": "Admin proposes a price",
        "description": "Admin submits a price proposal for a custom product. The product must be in `AwaitingAdminReview` status.\n\nAfter this call the status transitions to `AwaitingCustomerReview` — the customer will see the proposal and can accept or counter-offer.\n\nThe admin is also required to propose a price even if they agree with the system's automated quote. This is intentional: it forces an explicit review and ensures the admin is aware of the final number before the customer sees it.",
        "operationId": "CustomProduct_AdminPropose",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique identifier of the custom product",
            "required": true,
            "schema": { "type": "string" },
            "example": "cp-uuid-001"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProposeCustomPriceDto" },
              "example": {
                "amount": 6000,
                "comment": "We can offer this piece at 6000 given the resin material and complexity."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Price proposal registered. Status is now `AwaitingCustomerReview`.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CustomProductDto" },
                "example": {
                  "id": "cp-uuid-001",
                  "name": "Ring",
                  "description": "A custom engagement ring",
                  "price": 6000,
                  "taskId": "meshy-task-abc123",
                  "glbUrl": "https://cdn.meshy.ai/models/abc123.glb",
                  "ownerUserId": "user-uuid-001",
                  "status": "AwaitingCustomerReview",
                  "agreedPrice": null,
                  "design": {
                    "type": "Ring",
                    "material": "Resin",
                    "color": "Blue",
                    "shape": "Round",
                    "dimensions": "18x18x10",
                    "details": null
                  },
                  "proposals": [
                    {
                      "id": "proposal-uuid-001",
                      "amount": 7500,
                      "source": "System",
                      "comment": null,
                      "isAccepted": false,
                      "createdAt": "2026-05-10T14:00:00Z"
                    },
                    {
                      "id": "proposal-uuid-002",
                      "amount": 6000,
                      "source": "Admin",
                      "comment": "We can offer this piece at 6000 given the resin material and complexity.",
                      "isAccepted": false,
                      "createdAt": "2026-05-10T15:30:00Z"
                    }
                  ],
                  "createdAt": "2026-05-10T14:00:00Z",
                  "updatedAt": "2026-05-10T15:30:00Z"
                }
              }
            }
          },
          "400": {
            "description": "Product is not in `AwaitingAdminReview` state",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "The admin can not propose in this state" }
              }
            }
          },
          "401": { "description": "User is not authenticated" },
          "403": { "description": "User does not have the `Admin` role" },
          "404": { "description": "Custom product not found" }
        },
        "security": [{ "cookieAuth": [] }]
      }
    },
    "/api/CustomProduct/{id}/proposals": {
      "post": {
        "tags": ["CustomProduct - Customer"],
        "summary": "Customer counter-offers a price",
        "description": "Customer submits a counter-offer price. The product must be in `AwaitingCustomerReview` status and the authenticated user must be the product owner.\n\nAfter this call the status transitions back to `AwaitingAdminReview` — the admin will see the customer's counter and can accept or propose again.\n\n**Important:** Send the basket cookie with this request. The server captures the basket cookie to keep track of which cart should receive the product if admin later accepts this counter-offer.",
        "operationId": "CustomProduct_CustomerPropose",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique identifier of the custom product",
            "required": true,
            "schema": { "type": "string" },
            "example": "cp-uuid-001"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProposeCustomPriceDto" },
              "example": {
                "amount": 5200,
                "comment": "Would it be possible at 5200? That's my budget limit."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Counter-offer registered. Status is now `AwaitingAdminReview`.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CustomProductDto" },
                "example": {
                  "id": "cp-uuid-001",
                  "name": "Ring",
                  "description": "A custom engagement ring",
                  "price": 5200,
                  "taskId": "meshy-task-abc123",
                  "glbUrl": "https://cdn.meshy.ai/models/abc123.glb",
                  "ownerUserId": "user-uuid-001",
                  "status": "AwaitingAdminReview",
                  "agreedPrice": null,
                  "design": {
                    "type": "Ring",
                    "material": "Resin",
                    "color": "Blue",
                    "shape": "Round",
                    "dimensions": "18x18x10",
                    "details": null
                  },
                  "proposals": [
                    {
                      "id": "proposal-uuid-001",
                      "amount": 7500,
                      "source": "System",
                      "comment": null,
                      "isAccepted": false,
                      "createdAt": "2026-05-10T14:00:00Z"
                    },
                    {
                      "id": "proposal-uuid-002",
                      "amount": 6000,
                      "source": "Admin",
                      "comment": "We can offer this piece at 6000.",
                      "isAccepted": false,
                      "createdAt": "2026-05-10T15:30:00Z"
                    },
                    {
                      "id": "proposal-uuid-003",
                      "amount": 5200,
                      "source": "Customer",
                      "comment": "Would it be possible at 5200? That's my budget limit.",
                      "isAccepted": false,
                      "createdAt": "2026-05-10T16:00:00Z"
                    }
                  ],
                  "createdAt": "2026-05-10T14:00:00Z",
                  "updatedAt": "2026-05-10T16:00:00Z"
                }
              }
            }
          },
          "400": {
            "description": "Product is not in `AwaitingCustomerReview` state, or the user is not the product owner",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "The client can not propose in this state" }
              }
            }
          },
          "401": { "description": "User is not authenticated" },
          "404": { "description": "Custom product not found" }
        },
        "security": [{ "cookieAuth": [] }]
      }
    },
    "/api/CustomProduct/{id}/approve": {
      "post": {
        "tags": ["CustomProduct - Customer", "CustomProduct - Admin"],
        "summary": "Approve the current price proposal",
        "description": "Accepts the latest price proposal. Behavior depends on the caller's role:\n\n**Customer approves** (status must be `AwaitingCustomerReview`):\n- Validates that the basket cookie is present and the basket exists.\n- Marks the last proposal as accepted and sets `agreedPrice`.\n- Transitions status to `Approved`.\n- Automatically triggers addition to the cart via an integration event — **no further action needed from the frontend**. Poll `GET /api/Basket` to confirm the item appears.\n\n**Admin approves a customer counter-offer** (status must be `AwaitingAdminReview` and last proposal source is `Customer`):\n- Marks the last proposal as accepted and sets `agreedPrice`.\n- Transitions status to `Approved`.\n- Automatically triggers addition to the customer's cart using the basket they sent in their last interaction (submit or counter-offer). The frontend should notify the customer that the product was approved and added to their cart.\n\n**Preconditions for customer approval:**\n- The basket cookie must be present (the server reads it automatically from the request).\n- The basket identified by that cookie must exist in the system.\n\n**Preconditions for admin approval of customer counter-offer:**\n- The customer must have submitted or counter-offered at least once (so a `basketId` is on record). If they haven't, the server returns 400.",
        "operationId": "CustomProduct_Approve",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique identifier of the custom product",
            "required": true,
            "schema": { "type": "string" },
            "example": "cp-uuid-001"
          }
        ],
        "responses": {
          "200": {
            "description": "Price approved. Status is `Approved`. The product will be added to the cart asynchronously — poll `GET /api/Basket` after a short delay (1–2 seconds).",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CustomProductDto" },
                "example": {
                  "id": "cp-uuid-001",
                  "name": "Ring",
                  "description": "A custom engagement ring",
                  "price": 6000,
                  "taskId": "meshy-task-abc123",
                  "glbUrl": "https://cdn.meshy.ai/models/abc123.glb",
                  "ownerUserId": "user-uuid-001",
                  "status": "Approved",
                  "agreedPrice": 6000,
                  "design": {
                    "type": "Ring",
                    "material": "Resin",
                    "color": "Blue",
                    "shape": "Round",
                    "dimensions": "18x18x10",
                    "details": null
                  },
                  "proposals": [
                    {
                      "id": "proposal-uuid-001",
                      "amount": 7500,
                      "source": "System",
                      "comment": null,
                      "isAccepted": false,
                      "createdAt": "2026-05-10T14:00:00Z"
                    },
                    {
                      "id": "proposal-uuid-002",
                      "amount": 6000,
                      "source": "Admin",
                      "comment": "We can offer this piece at 6000.",
                      "isAccepted": true,
                      "createdAt": "2026-05-10T15:30:00Z"
                    }
                  ],
                  "createdAt": "2026-05-10T14:00:00Z",
                  "updatedAt": "2026-05-10T16:30:00Z"
                }
              }
            }
          },
          "400": {
            "description": "Cannot approve in current state, basket missing, or customer basket not on record for admin approval",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "invalidState": {
                    "summary": "Wrong state",
                    "value": { "error": "can not approve in this state" }
                  },
                  "basketRequired": {
                    "summary": "Customer basket cookie missing",
                    "value": { "error": "BasketId is required to approve" }
                  },
                  "unknownBasket": {
                    "summary": "Admin approving counter-offer but customer never sent a basket",
                    "value": { "error": "Customer basket is unknown; cannot complete approval" }
                  }
                }
              }
            }
          },
          "401": { "description": "User is not authenticated" },
          "403": { "description": "Customer trying to approve another user's product" },
          "404": {
            "description": "Custom product not found, or basket not found for customer approval",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "Basket not found" }
              }
            }
          }
        },
        "security": [{ "cookieAuth": [] }]
      }
    },
    "/api/CustomProduct/{id}/reject": {
      "post": {
        "tags": ["CustomProduct - Admin"],
        "summary": "Reject a custom product",
        "description": "Admin permanently rejects a custom product. The product transitions to `Rejected` status and cannot be approved or further negotiated.\n\nRejection is only allowed before the product reaches `Approved` or `AddedToBasket` status.",
        "operationId": "CustomProduct_Reject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique identifier of the custom product",
            "required": true,
            "schema": { "type": "string" },
            "example": "cp-uuid-001"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "description": "Reason for rejection — shown to the customer",
                "example": "The design complexity exceeds our current production capacity."
              },
              "example": "The design complexity exceeds our current production capacity."
            }
          }
        },
        "responses": {
          "200": {
            "description": "Custom product rejected successfully. Status is now `Rejected`."
          },
          "400": {
            "description": "Product is already `Approved` or `AddedToBasket` and cannot be rejected",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "can not reject a product approved" }
              }
            }
          },
          "401": { "description": "User is not authenticated" },
          "403": { "description": "User does not have the `Admin` role" },
          "404": { "description": "Custom product not found" }
        },
        "security": [{ "cookieAuth": [] }]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CustomProductDesignRequest": {
        "type": "object",
        "description": "Design parameters extracted from the Meshy AI generation result. These are used by the automated quotation engine to estimate the price.",
        "required": ["type", "material", "color", "shape", "dimensions"],
        "properties": {
          "type": {
            "type": "string",
            "description": "Category or type of the piece (e.g. Ring, Necklace, Bracelet, Pendant). Also used as the product name if no explicit name is provided.",
            "example": "Ring"
          },
          "material": {
            "type": "string",
            "description": "Primary material of the piece — impacts the quotation calculation",
            "example": "Resin"
          },
          "color": {
            "type": "string",
            "description": "Primary color of the piece",
            "example": "Blue"
          },
          "shape": {
            "type": "string",
            "description": "Overall shape of the piece",
            "example": "Round"
          },
          "dimensions": {
            "type": "string",
            "description": "Dimensions of the piece in millimeters as `widthxheightxdepth`. Used by the quotation engine to calculate volume and complexity.",
            "example": "18x18x10"
          },
          "details": {
            "type": "string",
            "nullable": true,
            "description": "Optional additional design details or finishing notes",
            "example": "Engraved with initials JD"
          }
        }
      },
      "ProposeCustomPriceDto": {
        "type": "object",
        "description": "Price proposal payload, used by both admin and customer",
        "required": ["amount"],
        "properties": {
          "amount": {
            "type": "integer",
            "format": "int64",
            "description": "Proposed price in the system's base unit (cents). For example, 6000 represents $60.00.",
            "minimum": 1,
            "example": 6000
          },
          "comment": {
            "type": "string",
            "nullable": true,
            "description": "Optional message to accompany the proposal. Visible to the other party.",
            "example": "We can offer this piece at a lower price given the material."
          }
        }
      },
      "CustomProductDto": {
        "type": "object",
        "description": "Full representation of a custom product, including its current status, design, price history, and negotiation proposals",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the custom product",
            "example": "cp-uuid-001"
          },
          "name": {
            "type": "string",
            "description": "Display name of the piece — defaults to the `design.type` value if not explicitly set",
            "example": "Ring"
          },
          "description": {
            "type": "string",
            "description": "Original description the customer provided to Meshy AI",
            "example": "A custom engagement ring with a blue gemstone"
          },
          "price": {
            "type": "integer",
            "format": "int64",
            "description": "Current working price in base units (cents). Reflects the last proposal amount. Once approved, matches `agreedPrice`.",
            "example": 6000
          },
          "taskId": {
            "type": "string",
            "description": "Meshy AI task identifier used to reference the 3D model",
            "example": "meshy-task-abc123"
          },
          "glbUrl": {
            "type": "string",
            "format": "uri",
            "description": "URL to the `.glb` 3D model. Use this to render the model in the frontend (e.g. with `<model-viewer>` or Three.js). Note: this is a `.glb` file, not an image.",
            "example": "https://cdn.meshy.ai/models/abc123.glb"
          },
          "ownerUserId": {
            "type": "string",
            "description": "User ID of the customer who owns this custom product",
            "example": "user-uuid-001"
          },
          "status": {
            "type": "string",
            "description": "Current status in the negotiation lifecycle",
            "enum": [
              "PendingQuotation",
              "AwaitingAdminReview",
              "AwaitingCustomerReview",
              "Approved",
              "Rejected",
              "AddedToBasket"
            ],
            "example": "AwaitingCustomerReview"
          },
          "agreedPrice": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "description": "Final agreed price in base units. `null` until both parties approve. Populated when status reaches `Approved` or `AddedToBasket`.",
            "example": 6000
          },
          "design": {
            "$ref": "#/components/schemas/CustomProductDesignRequest"
          },
          "proposals": {
            "type": "array",
            "description": "Chronological list of all price proposals. The last entry is always the active proposal. Use this array to render the negotiation history.",
            "items": { "$ref": "#/components/schemas/PriceProposalResponse" }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time when the product was submitted (UTC)",
            "example": "2026-05-10T14:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Date and time of the last status or price update (UTC). `null` if never updated after submission.",
            "example": "2026-05-10T16:30:00Z"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response returned by the API on 4xx responses",
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "Custom product not found"
          }
        }
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": ".AspNetCore.Identity.Application",
        "description": "ASP.NET Identity authentication cookie. Obtained via POST /api/login?useCookies=true."
      }
    }
  }
}